fix(scout-for-lol): resolve third-round Codex findings on /bb dare
Why: Codex's third review of the /bb dare PR found a batch-settlement
data-loss bug, a money-safety gap in dare confirmation, and a dead
mention in achievement copy — plus two file-size violations introduced
while fixing them.

What:
- settleDaresForMatch now collects every summary it successfully
  commits and throws once at the end via a new DarePartialSettlementError
  carrying all of them, instead of discarding already-committed
  results when a later dare in the same batch exhausts its retries.
  postmatch-hook.ts delivers those partial summaries before rethrowing.
- confirmDare renders the public callout and refuses with a new
  callout_too_long result before any money moves, so a dare can never
  debit a pledge it is then unable to announce.
- dareAchievedContent now actually mentions the challenger in the
  achieved-dare message body (previously only in allowedMentions.users,
  so the ping never rendered).
- Added orderBy: { id: "asc" } to the active-dare discovery query;
  Postgres doesn't guarantee findMany order, and a partial-batch test
  depends on deterministic processing order.
- Extracted dare-settle-shared.ts (types, DarePartialSettlementError,
  refund/void helpers) out of dare-settle.ts to stay under the 500-line
  cap, and split dare.integration.test.ts's largest describe block plus
  its storage-overflow/unreadable-conditions suites into a new
  dare-void.integration.test.ts to stay under the 1500-line test cap.
- Extracted dare-integration-fixtures.ts (shared constants, match
  builders, and a db-bound test-helper factory) so the two integration
  files reuse one implementation instead of duplicating it, keeping the
  jscpd duplication ratchet green after the split.

Verification:
- bunx tsc --noEmit -p packages/scout-for-lol/packages/backend
- bunx eslint (backend dare files, 0 errors)
- vitest src/betting/dare.integration.test.ts src/betting/dare-void.integration.test.ts (39 passed)
- bunx turbo run typecheck test lint --filter=@scout-for-lol/backend --filter=@scout-for-lol/data (20/20 passed)
- bun scripts/jscpd-check.ts (ratchet passed: 787 clones <= baseline 788)
