+++ packages/scout-for-lol/packages/backend/src/betting/dare-sweep.ts
@@ -12,7 +12,7 @@ import {
   type ActiveDareRow,
   type DareRefundView,
   type DareSettlementSummary,
-} from "#src/betting/dare-settle.ts";
+} from "#src/betting/dare-settle-shared.ts";
 import { logBucksTransition } from "#src/betting/transition-log.ts";
 import { prisma, type ExtendedPrismaClient } from "#src/database/index.ts";
 import { createLogger } from "#src/logger.ts";
